requirements.txt
aiohttp==3.9.3
python-socketio==5.11.1